# backend/DTOs/RegisterDTo.cs
namespace BlogAppApi.DTOs
{
public class RegisterDTo
{
public string Name { get; set; }

public string Email { get; set; }

public string Password { get; set; }

public string Role { get; set; } = "User";
}
}